feat: add support for specifying multiple docker-compose files
Docker-compose can read configuration from multiple docker files and
enabling an extension mechanism.

See: https://docs.docker.com/compose/extends/
